+++ source/src_test/abc_backlight_brightness_controller/fake_abc_io_service.c
@@ -40,7 +40,7 @@ abc_ioService_write(const int value, const char *const restrict pFileName)
 
     ++s_numWriteCalls;
 
-    s_currentBrightness = value;
+    s_currentBrightness = (uint16_t)value;
 
     s_isCurrentBrightnessSet = true;
